Themes: Render logged-out layout last
When logged-out, make sure that child components that go to `primary`,
`secondary`, and `tertiary` are first added to `layout`, before that latter is
rendered. This is achieved by modifying `LayoutLoggedOutDesign` to accept those
three as props, and by adding a new `makeLoggedOutLayout` middleware, which
adds the `LayoutLoggedOutDesign` layout to `context.layout`, which is rendered
by `clientRouter` (falling back to rendering `primary` and `secondary` as before).

This is necessary so we'll be able to use the same middleware on the server,
where the component we want to `renderToString()` needs to include all its
children when it's being rendered.